requirements.txt
google-analytics-data==0.10.0
google-api-core==1.31.5
google-auth==1.35.0
google-cloud==0.34.0
google-cloud-bigquery==2.32.0
google-cloud-bigquery-storage==2.0.0
google-cloud-storage==1.44.0
google-cloud-core==2.2.2
pandas==1.3.5
pandas-gbq==0.17.0